+++ services/121-service/src/metrics/dto/pa-metrics.dto.ts
@@ -11,8 +11,8 @@ export interface PaMetrics {
   [RegistrationStatusEnum.startedRegistration]: number;
   [RegistrationStatusEnum.registered]: number;
   [RegistrationStatusEnum.registeredWhileNoLongerEligible]?: number;
-  [RegistrationStatusEnum.selectedForValidation]?: number;
   [RegistrationStatusEnum.validated]: number;
+  [RegistrationStatusEnum.declined]: number;
   [RegistrationStatusEnum.included]: number;
   [RegistrationStatusEnum.inclusionEnded]: number;
   [RegistrationStatusEnum.rejected]: number;